+++ src/app/(front)/cricket/page.js
@@ -3,9 +3,12 @@
 import "./cricket.css";
 import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
+import { useSession } from "next-auth/react";
 
 export default function Cricketpage() {
   const router = useRouter();
+  const { data: session } = useSession();
+
   const [data, setData] = useState([]);
   const [err, setErr] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
@@ -33,6 +36,8 @@ export default function Cricketpage() {
 
   const handleDelete = async (id) => {
     try {
+      if (session?.user?.role !== "admin") return;
+
       await fetch(`/api/posts/${id}`, { method: "DELETE" });
       getData();
     } catch (err) {
@@ -61,7 +66,9 @@ export default function Cricketpage() {
 
       <div className="match-grid">
         {isLoading ? (
-          <div className="loader"><div className="spinner"></div></div>
+          <div className="loader">
+            <div className="spinner"></div>
+          </div>
         ) : err ? (
           <p className="error">Error loading posts.</p>
         ) : (
@@ -79,15 +86,11 @@ export default function Cricketpage() {
             )
             .map((post) => (
               <div className="card-wrapper" key={post._id}>
-
-
                 <div
                   className="match-card"
                   onClick={() => handleCardClick(post._id)}
-                  style={{ cursor: "pointer" }}
                 >
                   <div className="match-date">{post.date || "Today"}</div>
-
                   <div className="match-star">★</div>
 
                   <div className="match-flags image-bg">
@@ -100,12 +103,14 @@ export default function Cricketpage() {
                 </div>
 
 
-                <button
-                  className="delete-btn"
-                  onClick={() => handleDelete(post._id)}
-                >
-                  Delete
-                </button>
+                {session && session.user?.role === "admin" && (
+                  <button
+                    className="delete-btn"
+                    onClick={() => handleDelete(post._id)}
+                  >
+                    Delete
+                  </button>
+                )}
               </div>
             ))
         )}
